refactor: restructure the notifier package
