+++ internal/db/integration_types.go
@@ -1,0 +1,7 @@
+package db
+
+import "github.com/shinobistack/gokakashi/ent"
+
+func PopulateIntegrationTypes(client *ent.Client) {
+
+}
