+++ internal/integration/notification/notification.go
@@ -1,0 +1,13 @@
+package notification
+
+import (
+	"context"
+)
+
+type Notifier interface {
+	Notify(context.Context) error
+}
+
+type IntegrationType string
+
+var Linear IntegrationType = "linear"
