--- pkg/notifier/v1/notifier.go
@@ -1,17 +1,0 @@
-package notifier
-
-import (
-	"fmt"
-)
-
-func NewNotifier(tool string) (Notifier, error) {
-	switch tool {
-	case "linear":
-		return &LinearNotifier{}, nil
-	// Add other tools like Jira here
-	//case "jira":
-	//	return &JiraNotifier{}, nil
-	default:
-		return nil, fmt.Errorf("unsupported notifier tool: %s", tool)
-	}
-}
